Split statements into multiple files
